feat/water-2803 -- abstraction data integrated in to options -- added default values for loos based on abstraction data -- merged dev branch to get charge infor flow
