+++ src/internal/modules/charge-information/forms/charge-element/index.js
@@ -1,7 +1,7 @@
 module.exports = {
   purpose: require('./purpose'),
   description: require('./description'),
-  period: require('./abstraction-period'),
+  abstraction: require('./abstraction-period'),
   quantities: require('./quantities'),
   time: require('./time-limit'),
   source: require('./source'),
