+++ src/internal/modules/charge-information/forms/charge-element/purpose.js
@@ -3,12 +3,13 @@
 const urlJoin = require('url-join');
 const Joi = require('@hapi/joi');
 const { formFactory, fields } = require('shared/lib/forms/');
+const { uniqBy } = require('lodash');
 
-const options = [
-  { value: 'sid', label: 'Spray irrigation direct' },
-  { value: 'sido', label: 'Spray irrigation definition order' },
-  { value: 'sis', label: 'Spray irrigation storage' }
-];
+const options = defaultChargeData => {
+  return uniqBy(defaultChargeData.map(row => {
+    return { value: row.purposeUse.id, label: row.purposeUse.name };
+  }), 'value');
+};
 
 /**
  * Form to request if an FAO contact should be added to the invoice account
@@ -16,7 +17,7 @@ const options = [
  * @param {Object} request The Hapi request object
  * @param {Boolean}  selected value used to determine what radio option should be checked
   */
-const form = (request, sessionData = {}) => {
+const form = (request, sessionData = {}, defaultChargeData = []) => {
   const { csrfToken } = request.view;
   const { licenceId, elementId } = request.params;
   const action = urlJoin('/licences/', licenceId, 'charge-information/charge-element', elementId, 'purpose');
@@ -29,7 +30,7 @@ const form = (request, sessionData = {}) => {
         message: 'Select a purpose use'
       }
     },
-    choices: options
+    choices: options(defaultChargeData)
   }, sessionData.purpose || ''));
   f.fields.push(fields.hidden('csrf_token', {}, csrfToken));
   f.fields.push(fields.button(null, { label: 'Continue' }));
