+++ src/internal/modules/routes.js
@@ -16,7 +16,7 @@ const accountRoutes = require('./account/routes');
 const unlinkLicenceRoutes = require('./unlink-licence/routes');
 const chargingRoutes = require('./charging/routes');
 const billingRoutes = require('./billing/routes');
-const chargeInformationRoutes = require('./charge-information/routes');
+const chargeInformationRoutes = require('./charge-information/routes/charge-information');
 const invoiceAccRoutes = require('./invoice-accounts/routes');
 const addressEntryRoutes = require('./address-entry/routes');
