+++ test/internal/modules/charge-information/controller.js
@@ -16,7 +16,7 @@ const uuid = require('uuid/v4');
 const sandbox = sinon.createSandbox();
 
 const services = require('../../../../src/internal/lib/connectors/services');
-const controller = require('internal/modules/charge-information/controller');
+const controller = require('src/internal/modules/charge-information/controllers/charge-information');
 
 const createRequest = () => ({
   params: {
